requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
faiss-cpu==1.7.4
sentence-transformers==2.5.1
requests==2.31.0
python-dotenv==1.0.1
mistralai==0.0.12 